Redefine types
